+++ source/extensions/lightspeed.trex.project_wizard.core/docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [1.1.0] - 2023-09-05
+### Changed
+- Added extra checks to `is_project_file_valid`
+- Added more lenient rule for `_create_symlinks` existing project symlink in `rtx-remix` dir
+
 ## [1.0.0] - 2022-06-09
 ### Added
 - Created